scripts/requirements.txt
pandas
numpy
google-generativeai
geopy
fuzzywuzzy
python-Levenshtein